Stop using crcIds inside workflows.
The workflows can still used crcIds if desired (as legacy puzzle workflow)
inside them. The "public" facing methods which are designed for interaction
with the workflow still use crcIds as it doesn't make sense to pass different
copy of CaseRunConfiguration as argument since the workflow may have the
CaseRunConfiguration in different state so in this case, the reference (id)
is used.