Fix getting structures from original event
When a known structure is attempted to get from the run_subset event but the
structure is available only in original event (most of common cases), the
former code always returned None as there was no conversion possible in the
run_subset event and as AttributeError wasn't raised, the event didn't attempt
to get the structure from original event.

With this change, if the structure cannot be obtained from run_subset event
(None is returned), it's treated the same was as AttributeError and run_subset
attempts to get the structure from original event.