Introduce new builtin structure product.
The main purpose of this structure is to provide abstraction layer and common
interface for different structures to provide information about the tested
product in unified way so that it can be used for definition of desired
testplans repository branch.

Also drop direct passing of event structures to testplans repository branch
name formatting as this way prevents using the event structures conversion
which is now practically essential with this change.