Timeout only once when converting koji_build to compose
Since the conversion can be done from multiple different threads and
7e6a5cf94ee6d2e09ce2dea076fb53b3fc56edbd introduced conversion locking, it
could happen, that one thread would timeout on conversion and other thread
would wait again for the timeout (or the compose would appear in the meantime)
resulting in inconsistent behaviour. Introduction of this change should make
the settings option do exactly what it says.