Do not use implicit .txt suffix for automatically added logs.
New filename argument is introduced to provide ability for the workflows
to specify which filename should be used in the storage if the log doesn't
have file associated yet.

This change of API needs to be also reflected in GroupWorkflow.groupLog to
provide hint about filename with .txt suffix (as the groupLog opens the file
in text mode anyway).