Add support to for libpermian.plugins.* import to pylint
Thanks to this, there's no need to ignore any code. Also the pylint can be
called once for the base code as well as for all the plugins.

There's one big limitation with this approach and that pylint doesn't treat
the plugins as libpermian.plugins.PLUGIN_NAME but rather as PLUGIN_NAME which
can interfere with other python modules which have the same name as
PLUGIN_NAME.